hai-detection/requirements.txt
# HAI Detection Module Dependencies

# Core
python-dotenv>=1.0.0
requests>=2.28.0

# Database
# SQLite is built-in

# Optional: Clarity database access
# sqlalchemy>=2.0.0
# pyodbc>=4.0.0

# Optional: Ollama client (for LLM classification)
# httpx>=0.24.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
